smart_md_debugger/requirements.txt
# Python package dependencies:
markdown-it-py~=3.0.0
mdit_py_plugins~=0.4.0

# External dependencies (must be installed separately if certain features are used):
# - A LaTeX distribution (e.g., TeX Live, MiKTeX, MacTeX) for full PDF output simulation,
#   though this linter aims to reduce reliance on it for syntax checking.
#
# Note: The dependency on CLI pandoc is being replaced by markdown-it-py for linting.
